Clean/refactor notifier, add user validation
Fix a few notifier bugs, move reminder/user validator to their own
files, add Makefile to docs
